+++ src/unicode.h
@@ -21,7 +21,7 @@ extern size_t utf8len(char *s);
 extern size_t utf8len_start_stop(const char *start, const char *stop);
 extern int utf8charlen(char ch);
 extern int utf_dsplen(const char *s);
-extern int utf_string_dsplen(const char *s, size_t max_bytes);
+extern int utf_string_dsplen(const char *s, int max_bytes);
 extern int readline_utf_string_dsplen(const char *s, size_t max_bytes, size_t offset);
 extern const char *utf8_nstrstr(const char *haystack, const char *needle);
 extern const char *utf8_nstrstr_with_sizes(const char *haystack, int haystack_size, const char *needle, int needle_size);
